% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake.R
\docType{data}
\name{buffer}
\alias{buffer}
\title{Buffer of Lake Sunapee for \code{quickmapr} examples}
\format{SpatialPolygonsDataFrame}
\description{
This example data is a SpatialPolygonsDataFrame representing a buffer
of a single lake, Lake Sunapee.
}
\keyword{datasets}

