% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{build_map_data}
\alias{build_map_data}
\title{builds a map_data from many input types}
\usage{
build_map_data(...)
}
\arguments{
\item{...}{list, sp, or qmap objects}
}
\value{
list of spatial objects with names
}
\description{
builds a map_data from many input types
}
\keyword{internal}

