% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake.R
\docType{data}
\name{elev}
\alias{elev}
\title{Elevation Data for \code{quickmapr} examples}
\format{RasterLayer of 455 x 246 and 30 meter resolution}
\description{
This example data is a RasterLayer of the National 
Elevation Dataset for a buffer of Lake Sunapee.
}
\keyword{datasets}
