% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lake.R
\docType{data}
\name{examp_lakes}
\alias{examp_lakes}
\title{Lakes from the National Lakes Morphometry Dataset for \code{quickmapr}
examples}
\format{SpatialPolygonsDataFrame}
\description{
This example data is a polygon dataset of many lakes in New Hampshire.  Data
for each lake are from the first version of the
\href{https://edg.epa.gov/metadata/catalog/search/resource/details.page?uuid=\%7B7DF16BCC-F72D-422F-B5CE-4886AAC6109C\%7D}{National Lakes Morphometry
Dataset}.
}
\keyword{datasets}

