% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{f}
\alias{f}
\title{Returns qmap object to original extent}
\usage{
f(qmap_obj = NULL)
}
\arguments{
\item{qmap_obj}{a qmap_obj to restore}
}
\description{
Allows for restoring map to original extent.  The inital map is replayed and 
the map extent of the object is reset to match the current display.
}
\examples{
\dontrun{
data(lake)
x<-qmap(list(lake,buffer,elev))
zi(x)
f(x)
}
}
