% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{get_colors}
\alias{get_colors}
\title{gets color from input qmap}
\usage{
get_colors(...)
}
\arguments{
\item{...}{list, sp, or qmap objects}
}
\value{
vector of colors
}
\description{
gets color from input qmap
}
\keyword{internal}
