% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{get_range}
\alias{get_range}
\title{Gets the x and y diff of a qmap_obj}
\usage{
get_range(qmap_obj)
}
\arguments{
\item{qmap_obj}{an qmap object}
}
\value{
numeric vector indicating the size of the x and y extent
}
\description{
Gets the x and y diff of a qmap_obj
}
\keyword{internal}
