% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/i.R
\name{i}
\alias{i}
\title{Identify}
\usage{
i(spdata)
}
\arguments{
\item{spdata}{an \code{sp} object from which to identify features}
}
\value{
Returns a list that contains data for the selected object (data is
         NULL if not a Spatial DataFrame object), the \code{sp} object, and
         additional information for each object (e.g. area and perimeter for
         polygons).
}
\description{
Interactively select an \code{sp} or \code{raster} object and return the data
associated with it. \code{i()} only accepts a single input point.
}
\examples{
\dontrun{
data(lake)
qmap(list(lake,elev,samples))
i(lake)
i(samples)
i(elev)
}
}

