% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/l.R
\name{l}
\alias{l}
\title{Label features}
\usage{
l(spdata, spfield)
}
\arguments{
\item{spdata}{a sp object from which to take the labels}

\item{spfield}{a field in the spdata object to use to label the features}
}
\description{
It is useful to be able to provide labels to features on your map when
examining the results of given analysis.  This function adds labels for a
given layer.
}
\examples{
\dontrun{
data(lake)
qmap(list(lake,width))
l(lake,"COMID")
}
}

