% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lake.R
\docType{data}
\name{lake}
\alias{lake}
\title{Lake Data for \code{quickmapr} examples}
\format{SpatialPolygonDataframe with 1 lakes, each with 13 variables}
\description{
This example data is a SpatialPolygonsDataFrame
of a single lake, Lake Sunapee.
}
\keyword{datasets}

