% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lake.R
\docType{data}
\name{length}
\alias{length}
\title{Maximum Lake Length line for \code{quickmapr} examples}
\format{SpatialLines}
\description{
This example data is a line dataset representing the maximum lake length.
}
\keyword{datasets}

