% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{m}
\alias{m}
\title{Measure}
\usage{
m(qmap_obj = NULL)
}
\arguments{
\item{qmap_obj}{a \code{qmap} object from which to measure features.}
}
\description{
Allows interactive selection of points and returns length of selecte line in
units of the current \code{qmap} object.
}

