% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{make_jpw}
\alias{make_jpw}
\title{make jpeg world file}
\usage{
make_jpw(file, bbx, width)
}
\arguments{
\item{file}{output file name}

\item{bbx}{bounding box in map units}

\item{width}{width in pixels}
}
\description{
make jpeg world file
}
\keyword{internal}
