% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake.R
\docType{data}
\name{nlcd}
\alias{nlcd}
\title{Land Cover Data for \code{quickmapr} examples}
\format{RasterLayer of 467 x 262 and 30 meter resolution}
\description{
This example data is a RasterLayer of the National 
Land Cover Dataset for a buffer of Lake Sunapee.
}
\keyword{datasets}

