% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/p.R
\name{p}
\alias{p}
\title{Pan the current plot}
\usage{
p(...)
}
\arguments{
\item{...}{arguments to be passed to zoom::move.to.click.zoom(...)}
}
\description{
Interactively reposition the current plot. Click on location to center plot and
use ESC to quit.  This is a wrapper
function/shortcut for zoom::move.to.click.zoom(...).
}
\examples{
\dontrun{
data(lake)
qmap(list(lake,buffer,elev))
p()
}
}

