% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qmap.R
\name{plot.qmap}
\alias{plot.qmap}
\title{Default plotting of a qmap object}
\usage{
\method{plot}{qmap}(qmap_obj)
}
\arguments{
\item{qmap_obj}{input qmap class to plot}
}
\description{
Plots the qmap class and uses the order, colors, extent, and fill option
from \code{qmap}.
}

