% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qmap.R
\name{print.qmap}
\alias{print.qmap}
\title{Default printing of a qmap object}
\usage{
\method{print}{qmap}(qmap_obj)
}
\arguments{
\item{qmap_obj}{input qmap class to print}
}
\description{
Prints the summary of a qmap object
}

