% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qmap}
\alias{qmap}
\title{Display spatial data from a list.}
\usage{
qmap(mapdata, extent = NULL, order = 1:length(mapdata),
  colors = 1:length(mapdata))
}
\arguments{
\item{mapdata}{A list of spatial objects or a \code{qmap} object.}

\item{extent}{}

\item{order}{}

\item{colors}{}
}
\value{
Function displays a map from the input \code{mapdata} paramter
}
\description{
This funtion genarates a call to \code{ggplot2} and builds the quick map
visualization.
}
\examples{
data(lakes)
m(list(exampleLake,exampleElev))
}

