% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmapr.R
\name{quickmapr}
\alias{quickmapr}
\title{Quick visualization of spatial data with simple interactivity}
\description{
This package provides tools to quickly viusalize and interact with 
\code{\link{sp}} and \code{\link[raster]{raster}} objects.  Simple interactivity 
includes panning, zooming, identifying, and labelling.  The goal of 
\code{\link{quickmapr}} is to make working with spatial data easy and quick,
it is not intended to provide publication quality maps.
}

