% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{s}
\alias{s}
\title{Select}
\usage{
s(qmap_obj = NULL, s_idx = 1, loc = NULL)
}
\arguments{
\item{qmap_obj}{a \code{qmap} object from which to select features.}

\item{s_idx}{a numeric value specifying which data layer to select or a 
character string indicating the name of the layer. Defaults to 1.}

\item{loc}{A list with an x and y numeric indicating a location.  Default is 
to interactively get loc value.}
}
\value{
Returns a selected \code{sp} object
}
\description{
Interactively select a single \code{sp} or \code{raster} object and return the 
object. \code{s()} only accepts a single input point.
}
\examples{
\dontrun{
data(lake)
qm<-qmap(list(lake,elev,samples))
s(qm,"lake")
s(qm,3)
}
}
