% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap.R
\name{summary.qmap}
\alias{summary.qmap}
\title{Summary of qmap}
\usage{
\method{print}{qmap}(x, ...)
}
\arguments{
\item{x}{input qmap class to print}

\item{...}{options passed to summary}
}
\description{
Prvoides useful summary of a qmap object
}

