% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake.R
\docType{data}
\name{width}
\alias{width}
\title{Maximum lake width for \code{quickmapr} examples}
\format{SpatialLines}
\description{
This example data is a line dataset representing the maximum lake width.
Maximum lake width is defined as the longest line perpendciular to the
maximum lake length.
}
\keyword{datasets}

