% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ze}
\alias{ze}
\title{Zooms in on extent}
\usage{
ze(...)
}
\arguments{
\item{...}{arguments to be passed to zoom::out.zoom()}
}
\description{
This function uses the package \code{zoom} to provide zooming and panning functionality.
These are simply wrappers with shortened function names. This function zooms in on a
selected area.
}
\examples{
data(lake)
qmap(list(lake,buffer,elev)) \%>\% zi()
}

