% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zi.R
\name{zi}
\alias{zi}
\title{Zooms in on current plot}
\usage{
zi(...)
}
\arguments{
\item{...}{arguments to be passed to zoom::in.zoom(...)}
}
\description{
Interactively zoom in on the current plot. Click on location to zoom into and
use ESC to quit. This is a wrapper
function/shortcut for zoom::in.zoom(...).
}
\examples{
\dontrun{
data(lake)
qmap(list(lake,buffer,elev))
zi()
}
}

