% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zo.R
\name{zo}
\alias{zo}
\title{Zooms out on current plot}
\usage{
zo(...)
}
\arguments{
\item{...}{arguments to be passed to zoom::out.zoom(...)}
}
\description{
Interactively zoom out on the current plot. Click on location to zoom out from and
use ESC to quit. This is a wrapper function/shortcut for zoom::out.zoom(...).
}
\examples{
\dontrun{
data(lake)
qmap(list(lake,buffer,elev))
zo()
}
}

