% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{get_sp_type}
\alias{get_sp_type}
\title{Pull out essential info on sp class}
\usage{
get_sp_type(spdata)
}
\arguments{
\item{spdata}{an sp object}
}
\value{
character vector indicating point, line, polygon, or grid
}
\description{
Pull out essential info on sp class
}
\keyword{internal}

