% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/l.R
\name{l}
\alias{l}
\title{Label features}
\usage{
l(qmap_obj, lab_order = 1, field = NULL)
}
\arguments{
\item{qmap_obj}{a qmap object from which to pull the labels.  Raster layers
are ignored.  Will also accept \code{sp} objects.}

\item{lab_order}{identify which sp layer to label.  Defaults to first object
in qmap_obj$map_data.}

\item{field}{a field in the sp object to use to label the features.  Defaults
       to row.names().}
}
\description{
It is useful to be able to provide labels to features on your map when
examining the results of given analysis.  This function adds labels for a
given layer.
}
\examples{
\dontrun{
data(lake)
qmap(list(lake,width))
l(lake,'COMID')
}
}

