% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.R
\name{p}
\alias{p}
\title{Pan the current plot}
\usage{
p(qmap_obj = NULL, loc = NULL)
}
\arguments{
\item{qmap_obj}{A qmap object.  Optional, but performs better with larger 
data sets.}

\item{loc}{A list with an x and y numeric indicating a location.  Default is 
to interactively get loc value until escaped.}
}
\description{
Interactively reposition the current plot.   Works on an existing
 \code{qmap} object.  Simply pass that object to \code{p()}.  A single 
repositioning results and the extent of the \code{qmap} object is changed.
}
\examples{
\dontrun{
data(lake)
x<-qmap(list(lake,buffer,elev))
p()
## Or
p(x)
}
}
