% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/p.R
\name{p}
\alias{p}
\title{Pan the current plot}
\usage{
p(qmap_obj = NULL, ...)
}
\arguments{
\item{qmap_obj}{A qmap object.  Optional, but performs better with larger
data sets.}

\item{...}{arguments to be passed to zoom::move.to.click.zoom(...)}
}
\description{
Interactively reposition the current plot.  Requires a \code{qmap} object.
Each click repositions the map with the location clicked as the center.
The extent of the \code{qmap} object is changed.
}
\examples{
\dontrun{
data(lake)
x<-qmap(list(lake,buffer,elev))
p()
## Or
p(x)
}
}

