% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap.R
\name{plot.qmap}
\alias{plot.qmap}
\title{Default plotting of a qmap object}
\usage{
\method{plot}{qmap}(x, ...)
}
\arguments{
\item{x}{input qmap class to plotS}

\item{...}{options passed to image or plot}
}
\description{
Plots the qmap class and uses the order, colors, extent, and fill option 
from \code{qmap}.
}

