% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmap.R
\name{print.qmap}
\alias{print.qmap}
\title{Default plotting of a qmap object}
\usage{
\method{print}{qmap}(x, ...)
}
\arguments{
\item{x}{input qmap class to print}

\item{...}{options passed to plot}
}
\description{
Plots a qmap object
}

