% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qmap.R
\name{print.qmap}
\alias{print.qmap}
\title{Default printing of a qmap object}
\usage{
\method{print}{qmap}(x, ...)
}
\arguments{
\item{x}{input qmap class to print}

\item{...}{options passed to summary}
}
\description{
Prints the summary of a qmap object
}

