% Generated by roxygen2 (4.0.2): do not edit by hand
\name{qmap}
\alias{qmap}
\title{Display spatial data from a list.}
\usage{
qmap(mapdata, extent = NULL, order = 1:length(mapdata),
  colors = 1:length(mapdata), prj = TRUE)
}
\arguments{
\item{mapdata}{A list of spatial objects}

\item{extent}{A \code{sp} or \code{raster} object to use as the initial extent of the map.  Defaults
to the maximum extent of all input object}

\item{order}{draw order of the spatial object. Defaults to order in mapdata}

\item{colors}{line colors. Defaults to 1:length(mapdata)}

\item{prj}{Logical to check projections of input spatial objects.  Transformation, if needed, should be
done prior to mapping with \code{rgdal::spTransform()}.}
}
\value{
Function displays a map from the input \code{mapdata} paramter
}
\description{
This funtion genarates a call to \code{ggplot2} and builds the quick map
visualization.
}
\examples{
data(lake)
qmap(list(lake,buffer,elev))
}

