% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lake.R
\docType{data}
\name{samples}
\alias{samples}
\title{Point data for \code{quickmapr} examples}
\format{SpatialPointsDataFrame}
\description{
This is a fabricated point dataset for what might be a common type of data
collected for a lake.

This example data is a fabricated point dataset with data for each location.
}
\keyword{datasets}

