% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ze.R
\name{ze}
\alias{ze}
\title{Zooms in on extent}
\usage{
ze(...)
}
\arguments{
\item{...}{arguments to be passed to zoom::out.zoom()}
}
\description{
This function uses the package \code{zoom} to provide zooming and panning functionality.
These are simply wrappers with shortened function names. This function zooms in on a
selected area.
}
\examples{
data(lake)
qmap(list(lake,buffer,elev)) \%>\% zi()
}

