% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ze.R
\name{ze}
\alias{ze}
\title{Zooms in on extent}
\usage{
ze(qmap_obj = NULL, ...)
}
\arguments{
\item{qmap_obj}{A qmap object.  Optional, but performs better with larger
data sets.}

\item{...}{arguments to be passed to zoom::sq.zoom()}
}
\description{
Select a bounding box interactively and zoom to that extent. There are two
methods for this.  A purely interactive zoom where you simply click on two
location to specify the extent and use ESC to quit. This is a wrapper
function/shortcut for zoom::sq.zoom(...).  This works well with smaller data
sets; however it lags as the data increases in size.  The second method
performs better with larger datasets and requires you to create a \code{qmap}
object and pass that object to \code{zi()}.  A single zoom in to the selected
extent results and the extent of the \code{qmap} object is changed.
}
\examples{
\dontrun{
data(lake)
qmap(list(lake,buffer,elev))
ze()
}
}

