% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ze.R
\name{ze}
\alias{ze}
\title{Zooms in on extent}
\usage{
ze(...)
}
\arguments{
\item{...}{arguments to be passed to zoom::sq.zoom()}
}
\description{
Select a bounding box interactively and zoom to that extent. Click twice to
select corners of extent and ESC to quit.  This is a wrapper
function/shortcut for zoom::sq.zoom(...).
}
\examples{
\dontrun{
data(lake)
qmap(list(lake,buffer,elev))
ze()
}
}

