% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ze.R
\name{ze}
\alias{ze}
\title{Zooms in on extent}
\usage{
ze(qmap_obj = NULL)
}
\arguments{
\item{qmap_obj}{A qmap object.  Optional, but performs better with larger
data sets.}
}
\description{
Select a bounding box interactively and zoom to that extent. Works on an existing
 \code{qmap} object.  Simply pass that object to \code{ze()}.  A single zoom in to
 the selected
extent results and the extent of the \code{qmap} object is changed.
}
\examples{
\dontrun{
data(lake)
qmap(list(lake,buffer,elev))
ze()
}
}

