% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zi.R
\name{zi}
\alias{zi}
\title{Zooms in on current plot}
\usage{
zi(...)
}
\arguments{
\item{...}{arguments to be passed to zoom::in.zoom()}
}
\description{
This function uses the package \code{zoom} to provide zooming and panning functionality.
These are simply wrappers with shortened function names. This function zooms in by a set
amount.
}
\examples{
data(lake)
qmap(list(lake,buffer,elev)) \%>\% zi()
}

