% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zi.R
\name{zi}
\alias{zi}
\title{Zooms in on current plot}
\usage{
zi(qmap_obj = NULL, zoom_perc = 0.5, ...)
}
\arguments{
\item{qmap_obj}{A qmap object.  Optional, but performs better with larger
data sets.}

\item{zoom_perc}{A proportion to determine the zoom level.  The x and y
axes are reduced by this amount.  Default is 0.5.}

\item{...}{arguments to be passed to zoom::in.zoom(...)}
}
\description{
Interactively zoom in on the current plot. There are two methods for
zooming.  A purely interactive zoom where you simply click on location to
zoom into and use ESC to quit. This is a wrapper function/shortcut for
zoom::in.zoom(...).  This works well with smaller data sets; however it lags
as the data increases in size.  The second method performs better with
larger datasets and requires you to create a \code{qmap} object and pass
that object to \code{zi()}.  A single zoom in results and the extent of the
\code{qmap} object is changed.
}
\examples{
\dontrun{
data(lake)
x<-qmap(list(lake,buffer,elev))
zi()
##Or
zi(x)
}
}

