% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zi.R
\name{zi}
\alias{zi}
\title{Zooms in on current plot}
\usage{
zi(qmap_obj = NULL, zoom_perc = 0.5, ...)
}
\arguments{
\item{qmap_obj}{A qmap object.  Optional, but performs better with larger
data sets.}

\item{zoom_perc}{A proportion to determine the zoom level.  The x and y
axes are reduced by this amount.  Default is 0.5.}

\item{...}{arguments to be passed to zoom::in.zoom(...)}
}
\description{
Interactively zoom in on the current plot. Requires a
\code{qmap} object.  A single zoom in, with the location clicked at the
center, results and the extent of the \code{qmap} object is changed.
}
\examples{
\dontrun{
data(lake)
x<-qmap(list(lake,buffer,elev))
zi()
##Or
zi(x)
}
}

