% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zo.R
\name{zo}
\alias{zo}
\title{Zooms out on current plot}
\usage{
zo(qmap_obj = NULL, zoom_perc = 0.5, ...)
}
\arguments{
\item{qmap_obj}{A qmap object.  Optional, but performs better with larger
data sets.}

\item{zoom_perc}{A proportion to determine the zoom level.  The x and y
axes are increased by this amount.  Default is 0.5.}

\item{...}{arguments to be passed to zoom::out.zoom(...)}
}
\description{
Interactively zoom out on the current plot. There are two methods for
zooming.  A purely interactive zoom where you simply click on location to
zoom out from and use ESC to quit. This is a wrapper function/shortcut for
zoom::out.zoom(...).  This works well with smaller data sets; however it lags
as the data increases in size.  The second method performs better with
larger datasets and requires you to create a \code{qmap} object and pass
that object to \code{zo()}.  A single zoom in results and the extent of the
\code{qmap} object is changed.
}
\examples{
\dontrun{
data(lake)
qmap(list(lake,buffer,elev))
zo()
}
}

