% Generated by roxygen2 (4.0.2): do not edit by hand
\name{zo}
\alias{zo}
\title{Zooms out on current plot}
\usage{
zo(...)
}
\arguments{
\item{...}{arguments to be passed to zoom::out.zoom()}
}
\description{
This function uses the package \code{zoom} to provide zooming and panning functionality.
These are simply wrappers with shortened function names. This function zooms out by a set
amount.
}
\examples{
data(lake)
qmap(list(lake,buffer,elev)) \%>\% zi()
}

