% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zo.R
\name{zo}
\alias{zo}
\title{Zooms out on current plot}
\usage{
zo(qmap_obj = NULL, zoom_perc = 0.5, ...)
}
\arguments{
\item{qmap_obj}{A qmap object.  Optional, but performs better with larger
data sets.}

\item{zoom_perc}{A proportion to determine the zoom level.  The x and y
axes are increased by this amount.  Default is 0.5.}

\item{...}{arguments to be passed to zoom::out.zoom(...)}
}
\description{
Interactively zoom out on the current plot. Requires a
\code{qmap} object. A single zoom out results, with the location clicked at the
center, and the extent of the \code{qmap} object is changed.
}
\examples{
\dontrun{
data(lake)
qmap(list(lake,buffer,elev))
zo()
}
}

